requirements.txt
boto3==1.34.144

PyJWT==2.8.0
